third_party/WebKit/LayoutTests/inspector/timeline/timeline-window-filter.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script src="resources/timeline-data.js"></script>
<script>

function test()
{
    var timelineData = InspectorTest.timelineData();

    WebInspector.showPanel("timeline");
    var timeline = WebInspector.panels.timeline;

    var model = timeline._model;

    function createFileReader(file, delegate)
    {
        return new InspectorTest.FakeFileReader(timelineData, delegate, model.saveToFile.bind(model));
    }

    InspectorTest.override(model, "_createFileReader", createFileReader);

    model.loadFromFile({}, new WebInspector.Progress());
    timeline._currentViews[0]._refresh();

    function dumpRecordsCountForRange(windowLeft, windowRight)
    {
        var view = timeline._currentViews[0];
        timeline._overviewPane._overviewGrid.setWindow(windowLeft, windowRight);
        view._refresh();
        InspectorTest.addResult("mode = " + timeline._presentationModeSetting.get());
        InspectorTest.addResult("range = " + windowLeft + " - " + windowRight);
        InspectorTest.addResult("time range = " + timeline._windowFilter._windowStartTime + " - " + timeline._windowFilter._windowEndTime);
        InspectorTest.addResult("records count: " + view._presentationModel.filteredRecords().length);
        InspectorTest.addResult("");
    }

    timeline._selectPresentationMode(WebInspector.TimelinePanel.Mode.Events);
    dumpRecordsCountForRange(0, 1);
    dumpRecordsCountForRange(0.25, 0.75);
    dumpRecordsCountForRange(0.33, 0.66);

    timeline._selectPresentationMode(WebInspector.TimelinePanel.Mode.Memory);
    dumpRecordsCountForRange(0, 1);
    dumpRecordsCountForRange(0.25, 0.75);
    dumpRecordsCountForRange(0.33, 0.66);

    timeline._selectPresentationMode(WebInspector.TimelinePanel.Mode.Frames);
    dumpRecordsCountForRange(0, 1);
    dumpRecordsCountForRange(0.25, 0.75);
    dumpRecordsCountForRange(0.33, 0.66);

    timeline._overviewPane._overviewGrid.setWindow(0.1, 0.9);

    InspectorTest.addResult("--------------------------------------------------------");
    InspectorTest.addResult("mode = " + timeline._presentationModeSetting.get());
    InspectorTest.addResult("time range = " + timeline._windowFilter._windowStartTime + " - " + timeline._windowFilter._windowEndTime);
    timeline._selectPresentationMode(WebInspector.TimelinePanel.Mode.Events);
    InspectorTest.addResult("");

    InspectorTest.addResult("mode = " + timeline._presentationModeSetting.get());
    InspectorTest.addResult("time range = " + timeline._windowFilter._windowStartTime + " - " + timeline._windowFilter._windowEndTime);
    timeline._selectPresentationMode(WebInspector.TimelinePanel.Mode.Memory);
    
    InspectorTest.addResult("");

    InspectorTest.addResult("mode = " + timeline._presentationModeSetting.get());
    InspectorTest.addResult("time range = " + timeline._windowFilter._windowStartTime + " - " + timeline._windowFilter._windowEndTime);
    timeline._selectPresentationMode(WebInspector.TimelinePanel.Mode.Frames);

    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline window filter.
</p>

<p>
It applies different ranges to the OverviewGrid and expects that current view reflects the change.
</p>
<p>
Also it detects that after switching to another view it gets the proper time range.
</p>

</body>
</html>
